🪲 [Fix]: Fix retry logic (#20)
## Description

This pull request includes a small change to the `scripts/main.ps1`
file. The change modifies the loop conditions in the script to ensure
proper iteration and error handling.

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L35-R40):
Changed the loop initialization and condition from starting at 0 to
starting at 1 and adjusted the comparison to ensure the loop runs the
correct number of times.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas