🩹 [Patch]: Add debug-logging (#22)
## Description

This pull request includes several enhancements to the PowerShell
scripts to improve logging, error handling, and overall structure. The
most important changes are the addition of `begin`, `process`, and `end`
blocks, improved error handling, and enhanced logging for better
debugging.

### Enhancements to logging and structure:

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011R4-R9):
Added `begin`, `process`, and `end` blocks to provide a clearer
structure and added debug messages at the start and end of the script.
[[1]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011R4-R9)
[[2]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L75-R108)
*
[`scripts/outputs.ps1`](diffhunk://#diff-ee715ca93229232e95883bf00629fd14e3bf174cdc17b723c4cc5d70e6a60a58R4-L17):
Added `begin`, `process`, and `end` blocks to provide a clearer
structure and added debug messages at the start and end of the script.
[[1]](diffhunk://#diff-ee715ca93229232e95883bf00629fd14e3bf174cdc17b723c4cc5d70e6a60a58R4-L17)
[[2]](diffhunk://#diff-ee715ca93229232e95883bf00629fd14e3bf174cdc17b723c4cc5d70e6a60a58L30-R43)

### Improvements to error handling:

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L37-R46):
Enhanced error handling by adding `-ErrorAction Stop` to the
`Install-PSResource` command and logging warnings for exceptions.
*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L75-R108):
Wrapped the main process block in a `try` block to catch and rethrow
exceptions, ensuring any errors are properly handled and logged.
*
[`scripts/outputs.ps1`](diffhunk://#diff-ee715ca93229232e95883bf00629fd14e3bf174cdc17b723c4cc5d70e6a60a58L30-R43):
Wrapped the main process block in a `try` block to catch and rethrow
exceptions, ensuring any errors are properly handled and logged.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas