📝 [Docs]: Add reference to the `Template-Action` that uses `GitHub-Script` (#31)
## Description

This pull request includes a small change to the `README.md` file. The
change updates the description of the GitHub Action to better explain
its integration with the GitHub PowerShell module and provides a link to
create a new repository from a template.

- Fixes #29

Changes in documentation:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L3-R5):
Updated the description to clarify the integration with the GitHub
PowerShell module and added a link to create a new repository from the
PSModule template.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas