🩹 [Patch]: Add GH Auth status if not App context (#35)
## Description

This pull request includes an enhancement to the GitHub connection
logging process in the `scripts/info.ps1` file. The change ensures that
the GitHub CLI authentication status is checked and displayed if the
authentication type is not 'APP'.

Improvements to GitHub connection logging:

*
[`scripts/info.ps1`](diffhunk://#diff-82c586f67d16e32953b47a962c269d0a484f8aa660d71ad354e91fd2d4334cd9L28-R35):
Modified the logging process to store the GitHub context in a variable
and added a conditional check to display the GitHub CLI authentication
status when the authentication type is not 'APP'.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas