🩹 [Patch]: Add try/finally blocks (#38)
## Description

This pull request includes changes to the `action.yml` and the addition
of a new script file to improve the robustness and cleanliness of the
GitHub Action's execution.

Improvements to GitHub Action execution:

* `action.yml`: Modified the `run` block to include a `try` block for
executing scripts and a `finally` block to ensure the cleanup script is
always executed. This change enhances the robustness of the script
execution.

Addition of cleanup script:

* `scripts/clean.ps1`: Added a new PowerShell script to handle cleanup
tasks, including logging debug information and resetting environment
variables. This ensures that the environment is properly cleaned up
after the action runs.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas