🩹 [Patch]: Update environment variable prefixes for GitHub script inputs (#43)
## Description

This pull request focuses on updating environment variable names across
multiple scripts to ensure consistency and clarity. The changes involve
renaming variables from `GITHUB_ACTION_INPUT_*` to
`PSMODULE_GITHUB_SCRIPT_INPUT_*`. This is to reduce the change of
clobbering the envvars for the GitHub-Script action. Aligns with a
undocumented decision of having all inputs of actions and workflows as
`<OWNER>_<REPO_NAME>_INPUT_*`.

- Fixes #36 

As this is internals of the action, this is deemed a patch update as
nothing should rely on these variables apart from the scripts in the
action.

* Updated environment variable names in `action.yml` to use the
`PSMODULE_GITHUB_SCRIPT_INPUT_*` prefix.
* Modified `scripts/info.ps1` to reflect the new environment variable
names for `ShowInfo`, `Token`, and `Debug` settings.
[[1]](diffhunk://#diff-82c586f67d16e32953b47a962c269d0a484f8aa660d71ad354e91fd2d4334cd9L15-R16)
[[2]](diffhunk://#diff-82c586f67d16e32953b47a962c269d0a484f8aa660d71ad354e91fd2d4334cd9L31-R35)
[[3]](diffhunk://#diff-82c586f67d16e32953b47a962c269d0a484f8aa660d71ad354e91fd2d4334cd9L63-R64)
* Adjusted `scripts/init.ps1` to use the updated environment variable
names for `ShowInit`, `Version`, `Prerelease`, `Token`, `ClientID`, and
`PrivateKey`.
[[1]](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6L13-R24)
[[2]](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6L76-R78)
[[3]](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6L96-R98)
* Changed `scripts/outputs.ps1` to use the new environment variable name
for `ShowOutput`.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas