🩹 [Patch]: Update default values for WorkingDirectory to be '.' (#45)
## Description

This pull request includes updates to the default values for the
`WorkingDirectory` input in both the `README.md` and `action.yml` files.
These changes ensure consistency in the default working directory for
the script.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L24-R24):
Updated the default value for the `WorkingDirectory` input from `${{
github.workspace }}` to `'.'`.

Configuration updates:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L52-R52):
Updated the default value for the `WorkingDirectory` input from `${{
github.workspace }}` to `'.'`.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas