🩹 [Patch]: Update Action name to be that of the `Name` input (#48)
## Description

This pull request includes updates to the `action.yml` file to improve
the flexibility and customization of the GitHub Action. The most
important changes are:

Customization improvements:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L12-R12):
Changed the default value of the `Name` input from a string with quotes
to a plain string to ensure consistency.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L66-R66):
Updated the step name to use the `Name` input dynamically, allowing for
more customizable action names.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas