🚀 [Feature]: Configure PowerShell ErrorView settings (#49)
## 🚀 New Feature: `ErrorView` Configuration

This release introduces a new feature to configure the PowerShell
`$ErrorView` variable in the GitHub Action. This also sets the default
to `NormalView` whereas the default in PowerShell is `ConciseView`.
`NormalView` provides more information that would be great to get when
troubleshooting errors in GitHub workflows. Users can still override
this by adding their own settings in the provided script or using the
input to set a different value when building actions or workflows.

#### Details

* Introduced the `ErrorView` input in `action.yml`, including its
description, default value (`'NormalView'`), and marking it as optional.
* The input is mapped to the new environment variable
`PSMODULE_GITHUB_SCRIPT_INPUT_ErrorView` in the `action.yml` file.
* The environment variable is validated using wildcard matching against
a predefined list of valid views, applies the matched view setting.
* Added a description for the new `ErrorView` option in the `README.md`
file. This option allows users to configure the `$ErrorView` variable
using full or partial names of valid views.