🩹 [Patch]: Updated the org level login test with quotes (#52)
## Description

This pull request updates the `TestWorkflow.yml` file to test quoted
inputs for GitHub App secrets and adjusts the job name to reflect this
change.

Workflow updates:

*
[`.github/workflows/TestWorkflow.yml`](diffhunk://#diff-242a265d6d6bfff6094c9285345022d0e6d7ddde58504dfc80249fafbd89ba2cL472-R472):
Updated the `name` field of the `ActionTestWithGitHubAppOrg` job to
include "quoted inputs" for clarity.
*
[`.github/workflows/TestWorkflow.yml`](diffhunk://#diff-242a265d6d6bfff6094c9285345022d0e6d7ddde58504dfc80249fafbd89ba2cL482-R483):
Modified the `ClientID` and `PrivateKey` inputs in the `Action-Test`
step to use quoted syntax for testing purposes.