🪲 [Fix]: Fix setting the `$ErrorView` setting via inputs (#53)
This pull request introduces changes related to the handling of the
`ErrorView` configuration in a GitHub Actions workflow. The changes
streamline the `ErrorView` initialization logic by removing redundant
code and adding validation to ensure `ErrorView` is set correctly during
runtime.

### Changes to `ErrorView` handling:

*
[`.github/workflows/TestWorkflow.yml`](diffhunk://#diff-242a265d6d6bfff6094c9285345022d0e6d7ddde58504dfc80249fafbd89ba2cR228-R233):
Added a validation step to confirm that `ErrorView` is set to
`'NormalView'` during the workflow execution. If not, an exception is
thrown.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R94):
Introduced the assignment of the `ErrorView` environment variable from
the input parameter for use in subsequent scripts.
*
[`scripts/init.ps1`](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6L7-L23):
Removed the redundant initialization and validation logic for
`ErrorView`, as this is now handled directly in the workflow and action
configuration.