🪲 [Fix]: Fix debug and verbose inputs (#61)
This pull request introduces updates to improve debug and verbose output
handling across the GitHub PowerShell-based action. The changes ensure
consistent configuration of debug and verbose preferences and enhance
clarity in documentation and workflow files.

### Debug and Verbose Output Handling Updates:
*
[`.github/workflows/TestWorkflow.yml`](diffhunk://#diff-242a265d6d6bfff6094c9285345022d0e6d7ddde58504dfc80249fafbd89ba2cL412-R412):
Added the `-Debug` parameter to the `Get-GitHubUser` command to enable
debug output during the workflow execution.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L27-R31):
Updated the descriptions for `Debug` and `Verbose` inputs to clarify
that they enable debug and verbose output for the entire action.
Additionally, configured `$DebugPreference` and `$VerbosePreference`
based on input values to set PowerShell preferences dynamically.
[[1]](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L27-R31)
[[2]](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R95-R96)
*
[`scripts/info.ps1`](diffhunk://#diff-82c586f67d16e32953b47a962c269d0a484f8aa660d71ad354e91fd2d4334cd9L63-L64):
Removed redundant `$DebugPreference` and `$VerbosePreference`
configuration from the `end` block, as these preferences are now set
globally in the action runner.
*
[`scripts/outputs.ps1`](diffhunk://#diff-ee715ca93229232e95883bf00629fd14e3bf174cdc17b723c4cc5d70e6a60a58L6-L7):
Removed hardcoded `$DebugPreference` and `$VerbosePreference` settings
to align with the new dynamic configuration approach.

### Documentation Updates:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L18-R19):
Improved descriptions for `Debug` and `Verbose` inputs to specify that
they enable output for the entire action, enhancing clarity for users.